// apps/admin/src/generated/api/schemas/userSessionListResponseDto.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * Common Admin API
 * API for the common admin starter template
 * OpenAPI spec version: 0.1.0
 */
import type { UserSessionResponseDto } from "./userSessionResponseDto";

export interface UserSessionListResponseDto {
  items: UserSessionResponseDto[];
  total: number;
  page: number;
  pageSize: number;
}